You get a modal, you get a modal, everyone gets a modal (but now separated from the map component)
