+++ app/src/containers/MapContainer.js
@@ -1,0 +1,58 @@
+import { connect } from 'react-redux';
+import MapContainer from 'components/MapContainer';
+import { initGoogleMaps, setZoom, setCenter, setSupportModalVisibility } from 'actions/map';
+import { setUrlWorkspaceId, saveWorkspace } from 'actions/workspace';
+import { toggleLayerVisibility } from 'layers/layersActions';
+import { clearPolygon } from 'report/reportActions';
+import { loadTimebarChartData } from 'timebar/timebarActions';
+import { TIMELINE_OVERALL_START_DATE, TIMELINE_OVERALL_END_DATE } from 'constants';
+import { trackExternalLinkClicked } from 'analytics/analyticsActions';
+import { openShareModal, setShareModalError } from 'share/shareActions';
+
+const mapStateToProps = state => ({
+  activeBasemap: state.map.activeBasemap,
+  areas: state.areas.data,
+  basemaps: state.basemap.basemaps,
+  centerLat: state.map.center[0],
+  centerLong: state.map.center[1],
+  drawing: state.map.drawing,
+  layerIdPromptedForRemoval: state.layers.layerIdPromptedForRemoval,
+  maxZoom: state.map.maxZoom,
+  showMapCursorPointer: state.heatmap.highlightedVessels.isEmpty !== true && state.heatmap.highlightedVessels.clickableCluster !== true,
+  showMapCursorZoom: state.heatmap.highlightedVessels.clickableCluster === true,
+  token: state.user.token,
+  trackBounds: state.vesselInfo.trackBounds,
+  userPermissions: state.user.userPermissions,
+  zoom: state.map.zoom
+});
+
+const mapDispatchToProps = (dispatch, ownProps) => ({
+  initMap: (googleMaps) => {
+    dispatch(initGoogleMaps(googleMaps));
+  },
+  loadInitialState: () => {
+    dispatch(setUrlWorkspaceId(ownProps.workspaceId));
+    dispatch(loadTimebarChartData(TIMELINE_OVERALL_START_DATE, TIMELINE_OVERALL_END_DATE));
+  },
+  toggleLayerVisibility: (layer) => {
+    dispatch(toggleLayerVisibility(layer));
+  },
+  setZoom: zoom => dispatch(setZoom(zoom)),
+  setCenter: (center, centerWorld) => dispatch(setCenter(center, centerWorld)),
+
+  openShareModal: () => {
+    dispatch(openShareModal(true));
+    dispatch(saveWorkspace(setShareModalError));
+  },
+  openSupportModal: () => {
+    dispatch(setSupportModalVisibility(true));
+  },
+  clearReportPolygon: () => {
+    dispatch(clearPolygon());
+  },
+  onExternalLink: (link) => {
+    dispatch(trackExternalLinkClicked(link));
+  }
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(MapContainer);
