+++ config/webpack.config.js
@@ -74,21 +74,22 @@ const webpackConfig = {
       'node_modules'
     ],
     alias: {
-      assets: 'assets',
       actions: path.join(rootPath, 'app/src/actions'),
+      analytics: path.join(rootPath, 'app/src/analytics'),
+      app: path.join(rootPath, 'app'),
       areasOfInterest: path.join(rootPath, 'app/src/areasOfInterest'),
-      filters: path.join(rootPath, 'app/src/filters'),
-      search: path.join(rootPath, 'app/src/search'),
-      layers: path.join(rootPath, 'app/src/layers'),
-      report: path.join(rootPath, 'app/src/report'),
+      assets: 'assets',
       basemap: path.join(rootPath, 'app/src/basemap'),
+      filters: path.join(rootPath, 'app/src/filters'),
+      layers: path.join(rootPath, 'app/src/layers'),
+      lib: path.join(rootPath, 'app/lib'),
+      recentVessels: path.join(rootPath, 'app/src/recentVessels'),
+      report: path.join(rootPath, 'app/src/report'),
+      search: path.join(rootPath, 'app/src/search'),
+      share: path.join(rootPath, 'app/src/share'),
       siteNav: path.join(rootPath, 'app/src/siteNav'),
       timebar: path.join(rootPath, 'app/src/timebar'),
-      analytics: path.join(rootPath, 'app/src/analytics'),
       user: path.join(rootPath, 'app/src/user'),
-      recentVessels: path.join(rootPath, 'app/src/recentVessels'),
-      share: path.join(rootPath, 'app/src/share'),
-      lib: path.join(rootPath, 'app/lib'),
       components: 'src/components',
       constants: 'src/constants.js',
       containers: 'src/containers',
